python/requirements.txt
azure-functions
datadog-serverless-compat
ddtrace
